added spinner to explore view. Moved UI update of event cell completely over to dataManager. Added settings page link if user declines location services. Also performed general cleanup.

+++ PopulistMVP/PopulistMVP/PopulistMVP/PPLSTUUID.m
@@ -136,7 +136,7 @@ static CFMutableDictionaryRef CreateKeychainQueryDictionary(void)
 	
 	if (status != noErr)
 	{
-		NSLog(@"BPXLUUIDHandler Keychain Save Error: %ld", status);
+		NSLog(@"BPXLUUIDHandler Keychain Save Error: %ld", (long)status);
 		uuid = nil;
 	}
